Added b4g4r4a4 pixel format to read other textures
Was originally missing b4g4r4a4 pixel format, now properly reads as the 4th case was actually supposed to be reading for it after all. Left the old version in commented form, in case there's another reason why the same pixel format as 1 is used for img_type == 4 in certain edge cases.