Added img_type 6 (r8g8b8a8) and placeholders
After checking through all the files again, a single texture uses this minuscule format (8888 = 24 + 8). Seems the last commit was not the last type needed after all.
Then two more instances popped up. Image types 3 and 5 do not seem to be compatible with any option in raw texture cooker. Type 3 has a placeholder according to the closest existing option which may be right but I'm not sure. Type 5 is something that must be a custom RGBA format, though I'm not sure what it is. If anyone figures out the formats, send a PR for review please.