Added possible correct format for img_type 5
Updated image type 5 with the possible correct format for it. According to DDS image formats, bit masks are described in the files but running it through a raw image viewer and testing some values seems to have pulled the right ones. Problem is that the London textures "seem" right but I want to be 100% sure. Temporary fix for issue #3 until I can confirm without a doubt that the pixel format is correct.