Included support for "_nuno" (motion mesh) models
Loading the motion meshes was possible before. However, parenting and textures were incorrect. This commit adds a check to see if the model loaded is a motion mesh, then adjusts the parenting info accordingly instead of using the player parenting. Didi and Grinner have different parenting info in comparison to the other player models, therefore both have a different list of bone parenting to accommodate for. The other characters (JB, Tez, Miguel, Ripper) all use the same parent info, which is covered here. The script will also add the proper texture to the motion meshes with an added case for "_nuno".