Reworked script + manual input added
For the time being, manual input is being used for finding mesh start and faces start due to lack of pointers to such. I could not find them quite yet, so this is easier to at least get the format working for now. Code will be cleaned up when proper pointers are found. Some information lies below the header and first mesh start, but I'm not 100% on all of the details in it. Will work to research and improve mesh reading as time goes on. Produces output but incorrect output as of this moment. Covers Issue #2 partially, rework eschews loop entirely, and uses manual input to solve the issue temporarily.