Fixed stride values
The stride values previously were incorrect. The "padding" section for model researcher is actually a hex number (ex. 20 is 32 in decimal), so after correcting those proper output is achieved on limited models. The current limitation is due to the fact that some sections of mesh have submeshes, but the section information has the total amount of faces for all of the included submeshes. In the future, I'll need to find a way to separate them properly. Not stable yet, but functional for simple objects from the game (such as the goals). An Issue #2 milestone for proper output for sure.