Added prompt for loading all/specific animations
A dialogue box will pop up on selection of a character that asks the user to type 0 to select specific animations to load or type 1 to load all animations in the motion folder. This process will take a few seconds, so don't worry if it says not responding. It will finish after a bit.